Set up bolt 4.4 routing tests
The ROUTE message got changed in protocol version 4.4.
The new tests are guarded by the feature flag `BOLT_4_4 = "Feature:Bolt:4.4"`.
