+++ tests/stub/routing/scripts/v4x4/empty_reader.script
