+++ tests/stub/routing/scripts/v4x4/reader_tx.script
