+++ tests/stub/routing/scripts/v4x4/reader_tx_with_unexpected_interruption_on_run.script
