+++ tests/stub/routing/scripts/v4x4/reader_with_explicit_hello.script
