+++ tests/stub/routing/scripts/v4x4/router_connectivity_db.script
@@ -1,0 +1,14 @@
+!: BOLT #VERSION#
+
+C: HELLO {"scheme": "basic", "credentials": "c", "principal": "p", "user_agent": "007", "routing": #ROUTINGCTX#}
+S: SUCCESS {"server": "#SERVER_AGENT#", "connection_id": "bolt-123456789"}
+*: RESET
+{{
+    C: ROUTE #ROUTINGCTX# [] {}
+    S: SUCCESS { "rt": { "ttl": 1000, "db": "home-db", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": ["#HOST#:9010", "#HOST#:9011"], "role":"READ"}, {"addresses": ["#HOST#:9020", "#HOST#:9021"], "role":"WRITE"}]}}
+----
+    C: ROUTE #ROUTINGCTX# [] {"db": "system"}
+    S: SUCCESS { "rt": { "ttl": 1000, "db": "system", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": ["#HOST#:9010", "#HOST#:9011"], "role":"READ"}, {"addresses": ["#HOST#:9020", "#HOST#:9021"], "role":"WRITE"}]}}
+}}
+*: RESET
+?: GOODBYE
