+++ tests/stub/routing/scripts/v4x4/router_system_then_adb_with_bookmarks.script
@@ -1,0 +1,13 @@
+!: BOLT #VERSION#
+!: AUTO RESET
+!: ALLOW RESTART
+
+C: HELLO {"scheme": "basic", "credentials": "c", "principal": "p", "user_agent": "007", "routing": #ROUTINGCTX#}
+S: SUCCESS {"server": "#SERVER_AGENT#", "connection_id": "bolt-123456789"}
+{?
+    C: ROUTE #ROUTINGCTX# [] {"db": "system"}
+    S: SUCCESS { "rt": { "ttl": 1000, "db": "system", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": ["#HOST#:9000"], "role":"READ"}, {"addresses": ["#HOST#:9020"], "role":"WRITE"}]}}
+    ?: GOODBYE
+?}
+C: ROUTE #ROUTINGCTX# [ "SystemBookmark" ] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": ["#HOST#:9000"], "role":"READ"}, {"addresses": ["#HOST#:9020"], "role":"WRITE"}]}}
