+++ tests/stub/routing/scripts/v4x4/router_unreachable_db_then_adb.script
@@ -1,0 +1,13 @@
+!: BOLT #VERSION#
+!: AUTO RESET
+!: ALLOW RESTART
+
+C: HELLO {"scheme": "basic", "credentials": "c", "principal": "p", "user_agent": "007", "routing": #ROUTINGCTX#}
+S: SUCCESS {"server": "#SERVER_AGENT#", "connection_id": "bolt-123456789"}
+{?
+    C: ROUTE #ROUTINGCTX# [] {"db": "unreachable"}
+    S: SUCCESS { "rt": { "ttl": 1000, "db": "unreachable", "servers": []}}
+    ?: GOODBYE
+?}
+C: ROUTE #ROUTINGCTX# [] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": ["#HOST#:9010", "#HOST#:9011"], "role":"READ"}, {"addresses": ["#HOST#:9020", "#HOST#:9021"], "role":"WRITE"}]}}
