+++ tests/stub/routing/scripts/v4x4/router_with_leader_change.script
@@ -1,0 +1,20 @@
+!: BOLT #VERSION#
+
+A: HELLO {"{}": "*"}
+*: RESET
+C: ROUTE #ROUTINGCTX# [] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": ["#HOST#:9020"],"role": "WRITE"}, {"addresses": ["#HOST#:9006","#HOST#:9007"], "role": "READ"}, {"addresses": ["#HOST#:9000"], "role": "ROUTE"}]}}
+*: RESET
+?: GOODBYE
+C: ROUTE #ROUTINGCTX# [] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": ["#HOST#:9020"],"role": "WRITE"}, {"addresses": ["#HOST#:9006","#HOST#:9007"], "role": "READ"}, {"addresses": ["#HOST#:9000"], "role": "ROUTE"}]}}
+*: RESET
+?: GOODBYE
+C: ROUTE #ROUTINGCTX# [] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": [],"role": "WRITE"}, {"addresses": ["#HOST#:9006","#HOST#:9007"], "role": "READ"}, {"addresses": ["#HOST#:9000"], "role": "ROUTE"}]}}
+*: RESET
+?: GOODBYE
+C: ROUTE #ROUTINGCTX# [] {"db": "adb"}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "adb", "servers": [{"addresses": ["#HOST#:9021"],"role": "WRITE"}, {"addresses": ["#HOST#:9006","#HOST#:9007"], "role": "READ"}, {"addresses": ["#HOST#:9000"], "role": "ROUTE"}]}}
+*: RESET
+?: GOODBYE
