+++ tests/stub/routing/scripts/v4x4/router_yielding_no_readers_any_db.script
@@ -1,0 +1,7 @@
+!: BOLT #VERSION#
+!: AUTO RESET
+
+A: HELLO {"{}": "*"}
+C: ROUTE #ROUTINGCTX# [] {"[db]": {"U": "*"}}
+S: SUCCESS { "rt": { "ttl": 1000, "db": "any-db", "servers": [{"addresses": ["#HOST#:9000"], "role":"ROUTE"}, {"addresses": [], "role":"READ"}, {"addresses": ["#HOST#:9000"], "role":"WRITE"}]}}
+?: GOODBYE
