+++ tests/stub/routing/scripts/v4x4/router_yielding_procedure_not_found_failure_connectivity_db.script
@@ -1,0 +1,13 @@
+!: BOLT #VERSION#
+!: AUTO RESET
+
+C: HELLO {"scheme": "basic", "credentials": "c", "principal": "p", "user_agent": "007", "routing": #ROUTINGCTX#}
+S: SUCCESS {"server": "#SERVER_AGENT#", "connection_id": "bolt-123456789"}
+{{
+    C: ROUTE #ROUTINGCTX# [] {}
+----
+    C: ROUTE #ROUTINGCTX# [] {"db": "system"}
+}}
+S: FAILURE {"code": "Neo.ClientError.Procedure.ProcedureNotFound", "message": "blabla"}
+   IGNORED
+   <EXIT>
