+++ tests/stub/routing/scripts/v4x4/router_yielding_unknown_failure.script
@@ -1,0 +1,9 @@
+!: BOLT #VERSION#
+!: AUTO RESET
+
+C: HELLO {"scheme": "basic", "credentials": "c", "principal": "p", "user_agent": "007", "routing": #ROUTINGCTX#}
+S: SUCCESS {"server": "#SERVER_AGENT#", "connection_id": "bolt-123456789"}
+C: ROUTE #ROUTINGCTX# [] {"[db]": {"U": "*"}}
+S: FAILURE {"code": "Neo.ClientError.General.Unknown", "message": "wut!"}
+   IGNORED
+   <EXIT>
