+++ tests/stub/routing/scripts/v4x4/writer.script
