+++ tests/stub/routing/scripts/v4x4/writer_and_reader_tx_with_bookmark.script
