+++ tests/stub/routing/scripts/v4x4/writer_tx_with_leader_switch_and_retry.script
