+++ tests/stub/routing/scripts/v4x4/writer_tx_yielding_database_unavailable_failure.script
