+++ tests/stub/routing/scripts/v4x4/writer_tx_yielding_failure_on_run.script
