+++ tests/stub/routing/scripts/v4x4/writer_with_unexpected_interruption_on_pull.script
