+++ tests/stub/routing/test_routing_v3.py
@@ -1,10 +1,10 @@
 from nutkit.frontend import Driver
 import nutkit.protocol as types
 from ...shared import get_driver_name
-from .test_routing_v4x3 import RoutingV4x3
+from .test_routing_v4x4 import RoutingV4x4
 
 
-class RoutingV3(RoutingV4x3):
+class RoutingV3(RoutingV4x4):
     bolt_version = "3"
     server_agent = "Neo4j/3.5.0"
